third_party/WebKit/LayoutTests/css-parser/offset-parsing.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="resources/property-parsing-test.js"></script>
<script>
// Verifies that offset shorthand values are properly parsed

assert_valid_value("offset", "path('M 0 0 H 1') -200% auto");
assert_valid_value("offset", "none 50px reverse 30deg");
assert_invalid_value("offset", "path('M 0 0 H 1') reverse 30deg 50px");
assert_invalid_value("offset", "path('M 0 0 H 1') auto");
assert_invalid_value("offset", "none 30deg reverse");
assert_invalid_value("offset", "path('M 0 0 H 1') -200%");
assert_invalid_value("offset", "path('M 0 0 H 1') 50px");
assert_invalid_value("offset", "path('M 0 0 H 1')", "path('M 0 0 H 1') 0px auto");
assert_invalid_value("offset", "center center path('M 0 0 L 100 100 M 100 200 L 200 200 Z L 300 300 Z') 100% 90deg/left bottom");
assert_invalid_value("offset", "100px");
assert_invalid_value("offset", "100px none auto 90deg");
assert_invalid_value("offset", "auto");
assert_invalid_value("offset", "30deg");
assert_invalid_value("offset", "30deg path('M 20 30 A 60 70 80')");
assert_invalid_value("offset", "auto 30deg 90px");
assert_invalid_value("offset", "none /");
assert_invalid_value("offset", "none / 10px 20px 30deg");
assert_invalid_value("offset", "path('M 20 30 A 60 70 80') bottom");
</script>
